reorganize data under time data summary from software offline
