+++ lib/SoftwarePayload.py
@@ -3,6 +3,7 @@ import os
 from .SoftwareUtil import *
 from .SoftwareOffline import *
 from .SoftwareHttp import *
+from .TimeSummaryData import *
 
 # send the data that has been saved offline
 def sendOfflineData(isNewDay=False):
